naive implementation of a queue
Vector was used instead of a normal array, as the uni counterpart was implemented on java which makes all the
arrays heap allocated. Because of this i didn't implement a circular queue as trying to get a heap allocated
array was too difficult for me. This left some unused polymorhpism stuff.